Files Encryption
Encrypted the files due to a message from Kattis about having them
exposed to public: the message is as following : "If you do create a git
repo, please keep it private. There are quite a few people who like to
plagiarize solutions, which causes us problems. In the future, we may
start to penalize plagiarized solutions in the scoring. "